Add a GitHub Actions CI workflow.
More complicated than what I normally use, but this one ends up being a
bit quirky as we need to ensure that we have mod_perl built against the
Perl that we want to use for testing (and not the system-perl).

Worse yet, Ubuntu's Apache package is built with threads enabled, and
the Perl versions we are testing against are not.  Thus, also need to
rebuild Apache.

But wait, there's more!  Prior to Perl 5.20, it wasn't compiled out of
the box with shrplib, so we can't use those in our test matrix here
unless we want to recompile Perl as well (and no, I don't).

Lastly... Apache won't start if run as `root`, so we have to add our own
`github` User to the container in order for us to run tests via `sudo`.